Charge parser implementation class
